backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
fuzzywuzzy[speedup]
python-Levenshtein
